requirement.txt
numpy
torch
torchvision
tqdm
Pillow
scikit-learn
tensorflow
git+https://github.com/openai/CLIP.git
fairscale
git+https://github.com/tqth/recognize-anything.git